HW-5/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0
pydantic>=2.4
pydantic-settings
httpx
pymysql
python-dotenv
